Split SceneDocument into BranchOps + GroupOps mixins (slice 3)
Final slice of ADR-002 stage 2.2. Inside backend/domain/, the branch-tree
semantics (mark_branch_comparison_note/mark_branch_synthesis/
BRANCH_STATUS_VALUES/set_branch_status/set_final_deliverable/
_branch_parent_edge/_chat_subtree_ids/delete_chat_node/send_message/
chat_branch_history/get_branch_root/regenerate_response/
remove_associated_content_children/resolve_generate_image/
resolve_regenerate_image/collapse_branch) move to branches.py as the
BranchOps mixin, and the frame/container geometry engine
(_bbox_of_members through ungroup, plus _detach_node_from_membership)
moves to groups.py as GroupOps. graph.py keeps the dataclass fields (in
unchanged order - the generated __init__/__eq__ depend on it) plus the
core node/edge/chart/view-state methods, composing
`class SceneDocument(BranchOps, GroupOps)`.

Mixins over a strictly layered split, deliberately: the scoping pass
found four real core->groups upward calls (move_node/move_nodes/
remove_nodes/set_chat_collapsed all call _recompute_group_bounds or
_detach_node_from_membership) - a layered core with no knowledge of
groups would have needed a post-mutation hook, which is a behavior
change, not a move. Mixins share self, so every cross-call keeps its
exact current semantics. The mixins are field-less plain classes, so
@dataclass still harvests fields only from SceneDocument's own body -
generated methods are unchanged (verified below). BRANCH_STATUS_VALUES
becomes a BranchOps class attribute; both existing reference styles
(self. in set_branch_status, SceneDocument. in session_load.py) resolve
through the MRO unchanged.

Every moved block verified byte-identical against its extracted line
range; dataclass constructor signatures/field order/defaults diffed
against the pre-refactor baseline (identical); MRO + composition
asserted directly; full suite 1,187 passed.

GRAPHLINK_REPO_NAVIGATION.md's two canvas.py entries updated to the new
layout, including the patch-seam note (patch backend.domain.graph.<name>,
never backend.canvas.<name>, for document behavior).
