+++ backend/domain/graph.py
@@ -49,7 +49,9 @@ from graphlink_grid_view_settings import (
 from graphlink_navigation_pins import NavigationPinStore
 from graphlink_token_estimator import TokenEstimator
 
+from backend.domain.branches import BranchOps
 from backend.domain.content_codec import _content_codec
+from backend.domain.groups import GroupOps
 from backend.domain.model import (
     BRANCH_HORIZONTAL_SPACING,
     CHART_MAX_HEIGHT,
@@ -92,7 +94,7 @@ def _estimate_tokens(text: str) -> int:
 
 
 @dataclass
-class SceneDocument:
+class SceneDocument(BranchOps, GroupOps):
     """The canvas document for one session. Plain data + invariants; the
     R6 serializer will read/write exactly this shape."""
 
@@ -1242,413 +1244,6 @@ class SceneDocument:
             raise SceneError(f"unknown node: {node_id}")
         node.content = str(content)
 
-    def mark_branch_comparison_note(self, node_id: str, source_node_ids: list[str]) -> None:
-        """ADR-002 Workstream 1 ("Compare Branches"): stamps an already-created
-        note as the output of the Compare Branches agent and records which
-        branches it compared - called once, immediately after add_note +
-        set_note_content, mirroring set_group_color's own "extra setter call
-        right after creation" shape (see the WS intent wrapper in
-        register_canvas)."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "note":
-            raise SceneError(f"node is not a note node: {node_id}")
-        node.is_branch_comparison = True
-        node.item_ids = list(source_node_ids)
-
-    def mark_branch_synthesis(
-        self,
-        node_id: str,
-        source_node_ids: list[str],
-        instructions: str,
-        provider: str | None,
-        model: str | None,
-    ) -> None:
-        """ADR-002 Workstream 1 ("Synthesize Branches"): stamps an
-        already-created CHAT node as the output of the Synthesize Branches
-        agent - mirrors mark_branch_comparison_note's own "extra setter call
-        right after creation" shape, adapted for a chat-kind result instead
-        of a note-kind one (see SceneNode.is_branch_synthesis's own comment
-        for why this is a distinct method/flag rather than reusing Compare
-        Branches')."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "chat":
-            raise SceneError(f"node is not a chat node: {node_id}")
-        node.is_branch_synthesis = True
-        node.item_ids = list(source_node_ids)
-        node.synthesis_instructions = str(instructions)
-        node.provider = provider
-        node.model = model
-
-    #: ADR-002 Workstream 1 ("Branch status and lifecycle"): the exactly-4
-    #: legal values for SceneNode.branch_status - shared by the setter's
-    #: validation and session_load.py's own defensive downgrade-to-"active"
-    #: read-back, so the one legal set is never duplicated out of sync.
-    BRANCH_STATUS_VALUES = frozenset({"active", "accepted", "rejected", "superseded"})
-
-    def set_branch_status(self, node_id: str, status: str) -> None:
-        """ADR-002 Workstream 1 ("Branch status and lifecycle"): marks a
-        single chat node's own branch_status - deliberately no "has
-        siblings" / "is a fork root" requirement (any chat node may be
-        marked, mirroring mark_branch_comparison_note's own kind-only
-        validation), and deliberately no side effect on any OTHER node -
-        marking one branch Accepted does not auto-reject its siblings, the
-        first cross-node side-effecting setter would have been a new kind
-        of mutation nothing else in this file does, and Synthesize Branches
-        already established that 2+ branches can be simultaneously
-        legitimate (its own item_ids records multiple sources at once) -
-        forcing exclusivity here would fight that existing workflow."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "chat":
-            raise SceneError(f"node is not a chat node: {node_id}")
-        status = str(status)
-        if status not in self.BRANCH_STATUS_VALUES:
-            raise SceneError(f"invalid branch status: {status}")
-        node.branch_status = status
-
-    def set_final_deliverable(self, node_id: str, is_final: bool) -> None:
-        """ADR-002 Workstream 1 ("Branch status and lifecycle"): sets or
-        clears final_deliverable_node_id - EXCLUSIVE by construction (the
-        single-pointer shape means marking a new node silently supersedes
-        whichever one held it before; no separate "clear the old one" step
-        needed, unlike a per-node flag would require). Orthogonal to
-        branch_status on purpose - no validation ties them together (a
-        "rejected" node CAN technically be marked Final Deliverable; this
-        is not blocked, though not a realistic path either)."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "chat":
-            raise SceneError(f"node is not a chat node: {node_id}")
-        if is_final:
-            self.final_deliverable_node_id = node_id
-        elif self.final_deliverable_node_id == node_id:
-            self.final_deliverable_node_id = None
-
-    def _bbox_of_members(self, item_ids: list[str]) -> tuple[float, float, float, float]:
-        """Compute the padded union rect (x, y, width, height) enclosing
-        every member id's ESTIMATED footprint - GROUP_MEMBER_DEFAULT_WIDTH/
-        HEIGHT, since no current SceneNode kind carries its own width/height
-        field (see that constant's own comment). Stale/unknown member ids
-        (a member deleted out from under a group between mutations) are
-        silently skipped, never raise - a bbox recompute must never crash on
-        a dangling id. Falls back to a small default rect anchored at the
-        origin when item_ids is empty or every id is stale, so callers
-        (including resize_frame's own minimum-size clamp) always get a
-        well-defined rect back."""
-        left = top = right = bottom = None
-        for member_id in item_ids:
-            member = self.nodes.get(member_id)
-            if member is None:
-                continue
-            mx1, my1 = member.x, member.y
-            mx2 = member.x + GROUP_MEMBER_DEFAULT_WIDTH
-            my2 = member.y + GROUP_MEMBER_DEFAULT_HEIGHT
-            left = mx1 if left is None else min(left, mx1)
-            top = my1 if top is None else min(top, my1)
-            right = mx2 if right is None else max(right, mx2)
-            bottom = my2 if bottom is None else max(bottom, my2)
-        if left is None:
-            left = top = 0.0
-            right, bottom = GROUP_MEMBER_DEFAULT_WIDTH, GROUP_MEMBER_DEFAULT_HEIGHT
-        x = left - GROUP_PADDING
-        y = top - GROUP_PADDING_TOP
-        width = (right - left) + GROUP_PADDING * 2
-        height = (bottom - top) + GROUP_PADDING_TOP + GROUP_PADDING
-        return x, y, width, height
-
-    @staticmethod
-    def _union_rect(
-        a: tuple[float, float, float, float], b: tuple[float, float, float, float]
-    ) -> tuple[float, float, float, float]:
-        """The smallest (x, y, width, height) rect that fully contains both
-        inputs - legacy's own QRectF.united(), ported. The single primitive
-        _recompute_group_bounds uses to guarantee a frame's manual size
-        and/or manually-dragged position never clips a member: whichever
-        direction the live content has drifted, the result grows to cover
-        it, never shrinks below either input."""
-        ax, ay, aw, ah = a
-        bx, by, bw, bh = b
-        left = min(ax, bx)
-        top = min(ay, by)
-        right = max(ax + aw, bx + bw)
-        bottom = max(ay + ah, by + bh)
-        return left, top, right - left, bottom - top
-
-    def _recompute_group_bounds(self, node_id: str) -> None:
-        """The core "legacy never clips, always auto-grows to enclose
-        members" recompute - plain server-side math, NOT a React Flow
-        extent/parentId feature. Silent no-op for an unknown id or a
-        non-frame/container kind (defensive: every call site below already
-        only ever calls this with a live frame/container id, but a caller
-        that races a delete must never crash here).
-
-        Priority order:
-        1. Collapsed: skip the bbox computation entirely, snap to the fixed
-           GROUP_COLLAPSED_WIDTH/HEIGHT pill size. x/y are left untouched -
-           a collapsed pill stays wherever it was expanded from.
-        2. Frame with a manual size override and/or a manually-dragged
-           position (group_manual_width/height and/or group_manual_x/y
-           set): build a rect from whichever of those four are set (falling
-           back to the frame's current group_width/height for an unset
-           size, or the live bbox-of-members' own center for an unset
-           position), then UNION that rect with the live bbox-of-members -
-           never just substitute it. This is what makes both a manual
-           resize AND an independent drag stick (survive the very next
-           member move) without ever letting a member visually escape the
-           frame: if the live content has grown past the manual rect on any
-           edge, the union grows to re-enclose it instead of clipping or
-           silently reverting to bbox-centering.
-        3. Otherwise (auto-fit - every container, and every frame with
-           nothing manual set): x/y/width/height come straight from the
-           padded bbox-of-members.
-        """
-        node = self.nodes.get(node_id)
-        if node is None or node.kind not in ("frame", "container"):
-            return
-        if node.is_collapsed:
-            node.group_width = GROUP_COLLAPSED_WIDTH
-            node.group_height = GROUP_COLLAPSED_HEIGHT
-            return
-        bx, by, bw, bh = self._bbox_of_members(node.item_ids)
-        has_manual = node.kind == "frame" and (
-            node.group_manual_width is not None
-            or node.group_manual_height is not None
-            or node.group_manual_x is not None
-            or node.group_manual_y is not None
-        )
-        if has_manual:
-            width = node.group_manual_width if node.group_manual_width is not None else (node.group_width or bw)
-            height = node.group_manual_height if node.group_manual_height is not None else (node.group_height or bh)
-            if node.group_manual_x is not None and node.group_manual_y is not None:
-                anchor_x, anchor_y = node.group_manual_x, node.group_manual_y
-            else:
-                anchor_x = bx + bw / 2.0 - width / 2.0
-                anchor_y = by + bh / 2.0 - height / 2.0
-            node.x, node.y, node.group_width, node.group_height = self._union_rect(
-                (anchor_x, anchor_y, width, height), (bx, by, bw, bh)
-            )
-            return
-        node.x, node.y, node.group_width, node.group_height = bx, by, bw, bh
-
-    def _detach_from_existing_group(self, member_id: str, group_kind: str) -> None:
-        """Part of create_frame/create_container's shared validation: if
-        member_id is already tracked by some OTHER node of the SAME
-        group_kind ("frame" or "container" - membership is scoped per kind,
-        since a node may belong to at most one frame AND at most one
-        container simultaneously, per item_ids's own field comment), detach
-        it from that group first. If the detach empties that group's
-        item_ids, the now-empty group is deleted too (mirrors legacy
-        auto-delete-when-empty); otherwise the group's bounds are
-        recomputed to reflect its shrunk membership. A node can be a member
-        of at most one group of a given kind, so at most one match exists -
-        the loop stops at the first hit."""
-        for other in list(self.nodes.values()):
-            if other.kind != group_kind or member_id not in other.item_ids:
-                continue
-            other.item_ids = [i for i in other.item_ids if i != member_id]
-            if not other.item_ids:
-                self.nodes.pop(other.id, None)
-            else:
-                self._recompute_group_bounds(other.id)
-            break
-
-    def create_frame(self, item_ids: list[str]) -> SceneNode:
-        """Group an existing set of nodes into a new frame. Validates every
-        id exists AND is an eligible leaf-content kind BEFORE any mutation
-        (fail fast, no partial detach) - GROUP_INELIGIBLE_FRAME_MEMBER_KINDS
-        rejects a note or another frame/container, matching legacy's own
-        createFrame selection filter (frames never nest, and never absorb a
-        note - a note member would also be silently dropped from this
-        frame's own membership on save/reload, since frame_source_map has
-        no slot for one; see session_save.py). Then detaches each surviving
-        candidate from any frame it was already a member of (see
-        _detach_from_existing_group). is_locked defaults True (the legacy
-        frame default - locked). Initial x/y/width/height come from the
-        padded bbox-of-members, computed immediately via
-        _recompute_group_bounds right after construction."""
-        ids = list(item_ids)
-        for member_id in ids:
-            member = self.nodes.get(member_id)
-            if member is None:
-                raise SceneError(f"unknown member node: {member_id}")
-            if member.kind in GROUP_INELIGIBLE_FRAME_MEMBER_KINDS:
-                raise SceneError(
-                    f"node {member_id} (kind={member.kind!r}) cannot be a frame member - "
-                    f"frames only group leaf content nodes, never a note or another frame/container"
-                )
-        for member_id in ids:
-            self._detach_from_existing_group(member_id, "frame")
-        node_id = f"n{next(self._counter)}"
-        node = SceneNode(
-            id=node_id,
-            x=0.0,
-            y=0.0,
-            title="Frame",
-            kind="frame",
-            content="Add note...",
-            item_ids=ids,
-            is_locked=True,
-            is_collapsed=False,
-        )
-        self.nodes[node_id] = node
-        self._recompute_group_bounds(node_id)
-        return node
-
-    def create_container(self, item_ids: list[str]) -> SceneNode:
-        """Group an existing set of nodes into a new container. Same
-        validation/detach posture as create_frame, scoped to "container"
-        membership instead of "frame" - a node may simultaneously be a
-        member of one frame AND one container, so this never touches a
-        node's frame membership. UNLIKE create_frame, item_ids here may
-        include note/frame/container ids too - container membership can
-        nest (a container may hold another container or a frame as one of
-        its members). is_locked is left at its dataclass default (True) but
-        is MEANINGLESS for containers - see the field's own comment; no
-        toggle_container_lock exists and none should be added."""
-        ids = list(item_ids)
-        for member_id in ids:
-            if member_id not in self.nodes:
-                raise SceneError(f"unknown member node: {member_id}")
-        for member_id in ids:
-            self._detach_from_existing_group(member_id, "container")
-        node_id = f"n{next(self._counter)}"
-        node = SceneNode(
-            id=node_id,
-            x=0.0,
-            y=0.0,
-            title="Container",
-            kind="container",
-            content="New Container",
-            item_ids=ids,
-            is_collapsed=False,
-        )
-        self.nodes[node_id] = node
-        self._recompute_group_bounds(node_id)
-        return node
-
-    def set_group_label(self, node_id: str, text: str) -> None:
-        """Sets the header-note / title text for a frame or container -
-        reuses the generic `content` field, same reuse pattern as R3.5's
-        code text / R3.13's thinking text living in that same field."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind not in ("frame", "container"):
-            raise SceneError(f"node is not a frame/container node: {node_id}")
-        node.content = str(text)
-
-    def set_group_color(self, node_id: str, color: str | None, header_color: str | None) -> None:
-        """Shared color setter for note/frame/container kinds - see the
-        color/header_color fields' own comments on SceneNode for what each
-        controls. Either may be cleared back to None (default) by passing
-        None explicitly."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind not in ("note", "frame", "container"):
-            raise SceneError(f"node is not a note/frame/container node: {node_id}")
-        node.color = str(color) if color is not None else None
-        node.header_color = str(header_color) if header_color is not None else None
-
-    def toggle_frame_lock(self, node_id: str) -> None:
-        """Frame kind only. Recomputes bounds afterward for consistency with
-        every other group mutator here - locked vs unlocked does not change
-        the bbox math itself in this implementation (there is no
-        drag-suppression concept at the domain-model layer, only at the
-        frontend interaction layer), but keeping the call is cheap and
-        future-proofs a later change to that math."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "frame":
-            raise SceneError(f"node is not a frame node: {node_id}")
-        node.is_locked = not node.is_locked
-        self._recompute_group_bounds(node_id)
-
-    def toggle_group_collapsed(self, node_id: str) -> None:
-        """Shared frame/container collapse toggle. A single call to
-        _recompute_group_bounds after flipping is_collapsed correctly
-        handles BOTH directions: collapsing snaps to the fixed pill size
-        (that helper's own is_collapsed branch), expanding recomputes from
-        the bbox of members - respecting a frame's manual size override if
-        one is still set (group_manual_width/height survive the collapsed
-        state untouched, see those fields' own comments)."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind not in ("frame", "container"):
-            raise SceneError(f"node is not a frame/container node: {node_id}")
-        node.is_collapsed = not node.is_collapsed
-        self._recompute_group_bounds(node_id)
-
-    def resize_frame(self, node_id: str, width: float, height: float) -> None:
-        """Frame kind only. Records a manual size override, clamped to never
-        go below the padded bbox-of-members minimum size - computed via the
-        exact same _bbox_of_members helper the auto-fit path itself uses, so
-        "minimum" and "auto-fit size" can never drift apart. Recomputes
-        immediately afterward so x/y re-centers on the current member bbox
-        around the new size right away, same posture as toggle_frame_lock's
-        own trailing recompute call."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "frame":
-            raise SceneError(f"node is not a frame node: {node_id}")
-        _, _, min_width, min_height = self._bbox_of_members(node.item_ids)
-        node.group_manual_width = max(float(width), min_width)
-        node.group_manual_height = max(float(height), min_height)
-        self._recompute_group_bounds(node_id)
-
-    def fit_frame_to_content(self, node_id: str) -> None:
-        """Frame kind only. Clears BOTH the manual size override (set by
-        resize_frame) AND the manual position anchor (set by move_node
-        whenever this frame was dragged directly - see that method's own
-        comment) back to None, a full reset to pure auto-fit, then forces
-        an immediate bbox recompute. The size half is the exact inverse of
-        resize_frame; the position half is what makes this button also undo
-        an independent unlocked-frame drag, not just a resize."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "frame":
-            raise SceneError(f"node is not a frame node: {node_id}")
-        node.group_manual_width = None
-        node.group_manual_height = None
-        node.group_manual_x = None
-        node.group_manual_y = None
-        self._recompute_group_bounds(node_id)
-
-    def ungroup(self, node_id: str) -> None:
-        """Deletes a frame/container node itself. Members are NOT deleted
-        and keep their current absolute x/y positions unchanged - they
-        simply stop being tracked in any item_ids list (the deleted group's
-        own item_ids goes with it). Also drops any edges touching the group
-        node, mirroring remove_nodes' own "edges die with either endpoint"
-        invariant (frame/container nodes are not normally edge-connected the
-        way chat nodes are, but this keeps the invariant airtight rather
-        than relying on that never happening)."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind not in ("frame", "container"):
-            raise SceneError(f"node is not a frame/container node: {node_id}")
-        del self.nodes[node_id]
-        self.edges = {
-            eid: e for eid, e in self.edges.items()
-            if e.source != node_id and e.target != node_id
-        }
-        # A group can itself be a member of an outer container (nesting) -
-        # detach it there too, or the outer group is left tracking a
-        # dangling id, same failure mode remove_nodes already guards
-        # against for every other node kind.
-        self._detach_node_from_membership(node_id)
-
     # -- R6.2: chart node ----------------------------------------------------
 
     def add_chart_node(
@@ -1801,253 +1396,6 @@ class SceneDocument:
             raise SceneError(f"node is not a chart node: {node_id}")
         node.chart_aspect_locked = not node.chart_aspect_locked
 
-    def _branch_parent_edge(self, node_id: str) -> SceneEdge | None:
-        """R6.1: the shared 'find the edge whose target == node_id' lookup
-        chat_branch_history/get_branch_root/regenerate_response/
-        delete_chat_node all use to walk BRANCH structure (parent -> child,
-        source -> target) - factored out once this increment introduced a
-        second, unrelated edge shape that can also target a chat node: a
-        System Prompt note's note -> root edge (backend/plugins.py's
-        "System Prompt" branch, direction confirmed against
-        backend/agents.py's _resolve_branch_system_prompt). That edge is
-        METADATA (which note decorates this root) - a note is never a real
-        branch "parent", so a branch history/root walk must never traverse
-        through it (doing so would both corrupt chat_branch_history's real
-        conversation_history with the note's own content as a fake turn, AND
-        make get_branch_root resolve to the note instead of the true chat
-        root, silently defeating the override it exists to find). Skips
-        edges whose source is a kind="note" node for exactly that reason;
-        otherwise identical to the plain `next((e for e in self.edges.
-        values() if e.target == node_id), None)` pattern this replaces."""
-        for edge in self.edges.values():
-            if edge.target != node_id:
-                continue
-            source_node = self.nodes.get(edge.source)
-            if source_node is not None and source_node.kind == "note":
-                continue
-            return edge
-        return None
-
-    def _chat_subtree_ids(self, root_id: str) -> list[str]:
-        """ADR-002 Workstream 1 ("Branch status and lifecycle"): the
-        DOWNWARD counterpart to _branch_parent_edge's upward walk above -
-        every chat-kind node's id in root_id's own subtree (root_id
-        itself, plus every descendant reachable through chat-kind edges),
-        via BFS. No such downward/forward walk existed anywhere in this
-        file before this feature - every other edge scan here is either
-        this file's own upward parent-walk pattern or an explicitly
-        one-hop-only scan (delete_chat_node's direct-children reparent,
-        send_message's sibling-fan-out count) - so this is new, not a
-        rename of something existing. Chat-kind only: a code/document/
-        thinking/image node hanging directly off a chat node in this
-        subtree is NOT included - collapsing a branch does not cascade
-        into its content children, matching how a single chat node's own
-        collapse already never cascades into ITS children either."""
-        result = [root_id]
-        visited = {root_id}
-        frontier = [root_id]
-        while frontier:
-            next_frontier = []
-            for nid in frontier:
-                for edge in self.edges.values():
-                    if edge.source != nid:
-                        continue
-                    target = self.nodes.get(edge.target)
-                    if target is None or target.kind != "chat" or target.id in visited:
-                        continue
-                    visited.add(target.id)
-                    result.append(target.id)
-                    next_frontier.append(target.id)
-            frontier = next_frontier
-        return result
-
-    def delete_chat_node(self, node_id: str) -> None:
-        """Delete one chat node WITHOUT orphaning its branch: children are
-        re-parented to the deleted node's own parent (or become roots if it
-        had none), mirroring ChatScene.delete_chat_node's load-bearing
-        reparent rule - a plain remove_nodes cascade-delete would sever every
-        child branch instead of splicing them back together."""
-        if node_id not in self.nodes:
-            raise SceneError(f"unknown node: {node_id}")
-        parent_edge = self._branch_parent_edge(node_id)
-        parent_id = parent_edge.source if parent_edge is not None else None
-        child_edges = [e for e in self.edges.values() if e.source == node_id]
-        # R6.1: a System Prompt note attached to node_id (a note -> node_id
-        # edge - the exact shape _branch_parent_edge deliberately skips
-        # above, so it is NOT parent_edge) still dies with this endpoint,
-        # same "edges die with either endpoint" invariant remove_nodes
-        # already enforces elsewhere - otherwise it would dangle, pointing
-        # at a node_id that no longer exists in self.nodes. The note ITSELF
-        # is not deleted (mirrors ungroup's own "detach, don't
-        # cascade-delete" precedent) - only this now-stale edge.
-        note_edges = []
-        for edge in self.edges.values():
-            if edge.target != node_id:
-                continue
-            source_node = self.nodes.get(edge.source)
-            if source_node is not None and source_node.kind == "note":
-                note_edges.append(edge)
-
-        for edge in [parent_edge, *child_edges, *note_edges]:
-            if edge is not None:
-                self.edges.pop(edge.id, None)
-        if parent_id is not None:
-            for edge in child_edges:
-                self.connect(parent_id, edge.target)
-
-        if self.last_chat_node_id == node_id:
-            # The active branch continues from wherever it now ends: the
-            # deleted node's own parent (None if it had none either).
-            self.last_chat_node_id = parent_id
-
-        # ADR-002 Workstream 1 ("Branch status and lifecycle") - found by
-        # adversarial review: unlike last_chat_node_id above, this is
-        # cleared entirely rather than re-pointed to the parent. The parent
-        # was never itself marked Final Deliverable, so silently promoting
-        # it would misattribute a status the user never gave it; requiring
-        # a fresh, explicit re-mark is the safe behavior.
-        if self.final_deliverable_node_id == node_id:
-            self.final_deliverable_node_id = None
-
-        del self.nodes[node_id]
-        self._detach_node_from_membership(node_id)
-
-    def send_message(
-        self,
-        text: str,
-        content_parts: list[dict[str, Any]] | None = None,
-        branch_from_node_id: str | None = None,
-    ) -> SceneNode:
-        """The Composer's real Send action (R3.3): create a real user
-        ChatNode continuing the current branch (last_chat_node_id), or
-        start a fresh root if none exists yet. Positioning is a simple
-        deterministic stack, not the legacy find_branch_position packing
-        algorithm - real auto-layout is a later refinement; "Organize
-        Nodes" already exists as a fallback.
-
-        R8a: content_parts carries real attachments (image/audio) staged in
-        the composer - optional, additive, threaded straight to
-        add_chat_node.
-
-        ADR-002 Workstream 1 ("Branch from here"): branch_from_node_id, when
-        given and still a real node, OVERRIDES last_chat_node_id for this
-        one send - the actual fork primitive. Before this, last_chat_node_id
-        was the ONLY way to pick a parent, so a second real branch (two
-        children of one parent) was reachable only by manual edge
-        manipulation, never through the UI - see that field's own comment
-        ("until real node selection exists"). A bad/stale id (deleted node,
-        typo) falls through to the ordinary last_chat_node_id path rather
-        than raising, same defensive posture chat_branch_history's walk
-        already uses for an unknown id.
-
-        When branching onto a parent that already has one or more chat-kind
-        children (a genuine divergence, not a fresh continuation), the new
-        sibling fans out horizontally by BRANCH_HORIZONTAL_SPACING per
-        existing child instead of landing on the exact same (x, y) as an
-        existing branch - which would render as one node silently hiding
-        another.
-
-        last_chat_node_id is updated to the new node afterward exactly as
-        an ordinary send would be, override or not - so the branch just
-        created becomes the active one for the NEXT (non-overridden) send,
-        the same continue-from-here behavior as always."""
-        if branch_from_node_id is not None and branch_from_node_id in self.nodes:
-            parent_id: str | None = branch_from_node_id
-            parent = self.nodes[parent_id]
-            sibling_count = sum(
-                1
-                for e in self.edges.values()
-                if e.source == parent_id and (target := self.nodes.get(e.target)) is not None and target.kind == "chat"
-            )
-            x, y = parent.x + sibling_count * BRANCH_HORIZONTAL_SPACING, parent.y + MESSAGE_VERTICAL_SPACING
-        else:
-            parent_id = self.last_chat_node_id
-            if parent_id is not None and parent_id in self.nodes:
-                parent = self.nodes[parent_id]
-                x, y = parent.x, parent.y + MESSAGE_VERTICAL_SPACING
-            else:
-                parent_id = None
-                chat_node_count = sum(1 for n in self.nodes.values() if n.kind == "chat")
-                x, y = 0.0, chat_node_count * MESSAGE_VERTICAL_SPACING
-        node = self.add_chat_node(x, y, text, True, parent_id=parent_id, content_parts=content_parts)
-        self.last_chat_node_id = node.id
-        return node
-
-    def chat_branch_history(self, node_id: str) -> list[dict]:
-        """Walk the branch from node_id up to its root, collecting one
-        {"role", "content"} entry per node visited (including node_id
-        itself), then reverse so the result reads root-to-leaf (oldest
-        message first) - the direct new-backend replacement for legacy
-        conversation_history built by walking the QGraphicsScene parent
-        chain. Follows edges generically (by target match) rather than
-        asserting a chat-kind node shape: the walk itself only ever visits
-        chat-kind nodes in practice given how they are the only kind chained
-        this way, but a bad/unknown node_id or a stray edge shape should
-        stop the walk quietly rather than raise."""
-        history: list[dict] = []
-        current_id: str | None = node_id
-        while current_id is not None:
-            node = self.nodes.get(current_id)
-            if node is None:
-                break
-            # R8a: content_parts, when populated, IS the complete message
-            # content (its own text part plus any image_bytes/audio_file
-            # parts) - not an addendum to node.content. A node with no
-            # attachments has content_parts=None and this is byte-identical
-            # to before: a plain string, exactly as every other consumer of
-            # this history already expects.
-            history.append({
-                "role": "user" if node.is_user else "assistant",
-                "content": node.content_parts if node.content_parts else node.content,
-            })
-            parent_edge = self._branch_parent_edge(current_id)
-            current_id = parent_edge.source if parent_edge is not None else None
-        history.reverse()
-        return history
-
-    def get_branch_root(self, node_id: str) -> SceneNode | None:
-        """R6.1 addition (backend/agents.py's system-prompt-override
-        resolution and backend/plugins.py's System Prompt plugin both need
-        this same walk): find node_id's topmost ancestor by walking the
-        parent-edge chain up - the SAME by-target-match walk chat_branch_
-        history/regenerate_response/delete_chat_node already use (via
-        _branch_parent_edge) - until reaching a node with no incoming parent
-        edge. Returns node_id's own node when it already has no parent (it
-        IS the root), or None for an unknown node_id. Deliberately generic,
-        not scoped to kind="chat" - any node reachable via this edge-chain
-        shape can be walked."""
-        current_id: str | None = node_id
-        root: SceneNode | None = None
-        while current_id is not None:
-            node = self.nodes.get(current_id)
-            if node is None:
-                break
-            root = node
-            parent_edge = self._branch_parent_edge(current_id)
-            current_id = parent_edge.source if parent_edge is not None else None
-        return root
-
-    def regenerate_response(self, node_id: str) -> tuple[SceneNode, str]:
-        """Validate + resolve a regenerate target. Mirrors legacy's regenerate_node
-        single precondition (window_actions.py:512-514: no parent -> can't
-        regenerate), extended with two defensive checks legacy cannot hit (it
-        always holds a live scene-graph object, never a string id to resolve):
-        unknown node_id, and a non-chat-kind node_id (code/document/etc. can
-        never be regenerate targets directly - see Q2, the frontend always
-        resolves to a chat-node id before calling in). All three raise
-        SceneError; the WS-intent wrapper in register_canvas catches it and
-        shows ONE friendly notification for all three cases - see that wrapper
-        for why."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "chat":
-            raise SceneError(f"node is not a chat node: {node_id}")
-        parent_edge = self._branch_parent_edge(node_id)
-        if parent_edge is None:
-            raise SceneError(f"node has no parent and cannot be regenerated: {node_id}")
-        return node, parent_edge.source
-
     def update_chat_node_content(self, node_id: str, content: str) -> SceneNode:
         """The regenerate primitive: mutate an EXISTING chat node's content in
         place - the first in-place mutation of a content-bearing field in this
@@ -2063,80 +1411,6 @@ class SceneDocument:
             raise SceneError(f"unknown node: {node_id}")
         node.content = str(content)
         return node
-
-    def remove_associated_content_children(self, chat_node_id: str) -> None:
-        """The regenerate teardown: remove every code/document/image/thinking
-        node ONE HOP directly off chat_node_id. Built entirely on the existing
-        generic remove_nodes (edge cleanup + image-asset eviction come free).
-        Mirrors graphlink_scene.py's remove_associated_content_nodes exactly in
-        SCOPE (one-hop only, same four kinds, no cascade to any grandchild) but
-        resolved via this backend's one edge-encoded parent/child relationship
-        instead of legacy's four parallel per-kind lists. html/conversation
-        kinds are excluded on purpose - grep confirms neither ever has a
-        parent_content_node attribute in legacy, so they structurally can never
-        attach to a ChatNode this way."""
-        child_ids = []
-        for edge in self.edges.values():
-            if edge.source != chat_node_id:
-                continue
-            child = self.nodes.get(edge.target)
-            if child is not None and child.kind in ("code", "document", "image", "thinking"):
-                child_ids.append(child.id)
-        self.remove_nodes(child_ids)
-
-    def resolve_generate_image(self, chat_node_id: str) -> tuple[str, str]:
-        """'Generate Image from Text' target resolution (R4.4a). Returns
-        (parent_chat_node_id, prompt) = (chat_node_id, node.content) - the
-        selected ChatNode's own id becomes the new image's parent chat node,
-        and its own text becomes the prompt, mirroring legacy's real
-        "Generate Image from Text" entry point (window_actions.py's
-        generate_image(chat_node), called with node.text as the prompt).
-        Raises SceneError for an unknown node id or a non-chat kind
-        (defensive - the frontend always resolves this from a real ChatNode's
-        own menu, same posture as regenerate_response's own defensive checks
-        above), and the SceneEmptyPromptError subclass specifically for
-        empty/whitespace content - mirrors legacy's own "no text to use as a
-        prompt" guard (window_actions.py:989-991), kept as a DISTINCT
-        SceneError subclass (not a plain SceneError) so the WS wrapper in
-        register_canvas can show a distinct message for this case without
-        string-sniffing."""
-        node = self.nodes.get(chat_node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {chat_node_id}")
-        if node.kind != "chat":
-            raise SceneError(f"node is not a chat node: {chat_node_id}")
-        if not node.content or not node.content.strip():
-            raise SceneEmptyPromptError(f"node has no text to use as a prompt: {chat_node_id}")
-        return chat_node_id, node.content
-
-    def resolve_regenerate_image(self, image_node_id: str) -> tuple[str, str]:
-        """'Regenerate Image' target resolution (R4.4a). Returns
-        (parent_chat_node_id, prompt) = (the ImageNode's own parent chat node
-        id via one-hop edge lookup, node.content - the ImageNode's OWN stored
-        prompt). This is the deliberate improvement over legacy's real
-        regenerate mechanism, which instead re-derives the prompt from the
-        parent ChatNode's live .text - a real, reproducible legacy quirk that
-        re-wraps its own wrapped "Generated image for prompt: ..." string on
-        every subsequent regenerate. Raises SceneError for an unknown node
-        id, a non-image kind, or a missing parent edge (defensive only -
-        add_image_node requires parent_id, so an unparented image node can
-        never actually be constructed; this exists purely so a future bug
-        elsewhere fails loud instead of crashing downstream), and the
-        SceneEmptyPromptError subclass for empty/whitespace content
-        (defensive - mirrors legacy's own conditional-visibility guard `if
-        parent_content_node and prompt` around showing the menu action at
-        all)."""
-        node = self.nodes.get(image_node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {image_node_id}")
-        if node.kind != "image":
-            raise SceneError(f"node is not an image node: {image_node_id}")
-        parent_edge = self._branch_parent_edge(image_node_id)
-        if parent_edge is None:
-            raise SceneError(f"image node has no parent: {image_node_id}")
-        if not node.content or not node.content.strip():
-            raise SceneEmptyPromptError(f"image node has no prompt to regenerate from: {image_node_id}")
-        return parent_edge.source, node.content
 
     def add_generated_image_reply(
         self,
@@ -2186,31 +1460,6 @@ class SceneDocument:
         # it, rather than only being safe via toggle_group_collapsed.
         if node.kind in ("frame", "container"):
             self._recompute_group_bounds(node_id)
-
-    def collapse_branch(self, node_id: str, collapsed: bool) -> None:
-        """ADR-002 Workstream 1 ("Branch status and lifecycle"): "Collapse
-        a rejected branch without deleting it" - reuses the existing,
-        already-fully-wired is_collapsed field verbatim (wire sync, save,
-        load, and ChatNodeView.tsx's collapsed-pill rendering all already
-        work for it), applied across node_id's own chat-kind subtree via
-        _chat_subtree_ids instead of just node_id itself - the one
-        genuinely new piece this needs. Deliberately NOT automatic when a
-        branch is marked "rejected" via set_branch_status: status (a
-        semantic label) and collapse (a view state) are kept decoupled on
-        purpose, so a branch can be Rejected-but-still-expanded during
-        review, and marking status never has a side effect on any other
-        node's state (matching set_branch_status's own "no implicit side
-        effects" posture). Bulk-sets uniformly across the whole subtree,
-        so a node that had previously been individually expanded/collapsed
-        differently loses that distinction the first time this runs - an
-        accepted, stated tradeoff, not solved here."""
-        node = self.nodes.get(node_id)
-        if node is None:
-            raise SceneError(f"unknown node: {node_id}")
-        if node.kind != "chat":
-            raise SceneError(f"node is not a chat node: {node_id}")
-        for nid in self._chat_subtree_ids(node_id):
-            self.nodes[nid].is_collapsed = bool(collapsed)
 
     def set_all_conversational_collapsed(self, collapsed: bool) -> None:
         """R7.5e: Collapse All / Expand All - the bulk counterpart to
@@ -2344,32 +1593,6 @@ class SceneDocument:
                 if node.chart_asset_id:
                     self.image_assets.pop(node.chart_asset_id, None)
                 self._detach_node_from_membership(node_id)
-
-    def _detach_node_from_membership(self, node_id: str) -> None:
-        """R6.1: if the deleted node was itself a frame/container, its own
-        item_ids simply goes with it (already popped from self.nodes by the
-        caller) - members are NOT cascade-deleted, they just stop being
-        tracked, same "release, don't destroy" rule ungroup() uses. If the
-        deleted node was instead a MEMBER of some other frame/container (or,
-        since containers can nest, a group nested inside another group),
-        detach it from that group's item_ids - auto-deleting the group if
-        that empties it out (mirrors create_frame/create_container's own
-        detach rule), else recomputing its bounds to reflect the shrunk
-        membership. Shared by remove_nodes() and delete_chat_node() - the
-        latter deletes via its own reparent-children path rather than
-        remove_nodes, so it would otherwise leave stale item_ids behind.
-        list(...) over a live view since a match can mutate self.nodes
-        (popping an emptied group) mid-iteration."""
-        for group in list(self.nodes.values()):
-            if group.kind not in ("frame", "container"):
-                continue
-            if node_id not in group.item_ids:
-                continue
-            group.item_ids = [i for i in group.item_ids if i != node_id]
-            if not group.item_ids:
-                self.nodes.pop(group.id, None)
-            else:
-                self._recompute_group_bounds(group.id)
 
     # -- edges -------------------------------------------------------------
